+++ libs/angular-accelerator/src/lib/model/data-action.ts
@@ -5,5 +5,8 @@ export interface DataAction {
   permission: string
   classes?: string[]
   disabled?: boolean
+  actionVisibleField?: string
+  actionEnabledField?: string
+  showAsOverflow?: boolean
   callback: (data: any) => void
 }