+++ libs/angular-webcomponents/src/index.ts
@@ -1,0 +1,3 @@
+export * from './lib/utils/webcomponent-bootstrap.utils'
+export * from './lib/utils/webcomponent-router-initializer.utils'
+export * from './lib/utils/webcomponent-router.utils'