+++ libs/portal-integration-angular/src/lib/services/portal-dialog.service.ts
@@ -168,7 +168,7 @@ export interface DialogSecondaryButtonDisabled {
  * ```
  */
 export interface DialogButtonClicked<T = unknown> {
-  ocxDialogButtonClicked(state: DialogState<T>): Observable<boolean> | Promise<boolean> | boolean | undefined
+  ocxDialogButtonClicked(state: DialogState<T>): Observable<boolean> | Promise<boolean> | boolean | undefined | void
 }
 
 /**
@@ -206,6 +206,31 @@ type Component<T extends unknown> = unknown extends T
 export type DialogState<T> = {
   button: 'primary' | 'secondary'
   result: T | undefined
+}
+
+export type PortalDialogConfig = {
+  showXButton?: boolean
+  ariaLabelledBy?: string
+  width?: string
+  height?: string
+  closeOnEscape?: boolean
+  focusOnShow?: boolean
+  focusTrap?: boolean
+  baseZIndex?: number
+  autoZIndex?: boolean
+  dismissableMask?: boolean
+  showHeader?: boolean
+  modal?: boolean
+  resizable?: boolean
+  draggable?: boolean
+  keepInViewport?: boolean
+  minX?: number
+  minY?: number
+  maximizable?: boolean
+  maximizeIcon?: string
+  minimizeIcon?: string
+  position?: string
+  closeAriaLabel?: string
 }
 
 @Injectable({ providedIn: 'any' })
@@ -366,8 +391,31 @@ export class PortalDialogService {
     componentOrMessage: Type<any> | Type<DialogResult<T>> | Component<T> | TranslationKey | DialogMessage,
     primaryButtonTranslationKeyOrDetails: TranslationKey | ButtonDialogButtonDetails,
     secondaryButtonTranslationKeyOrDetails?: TranslationKey | ButtonDialogButtonDetails,
-    showXButton: boolean = true
+    extras?: PortalDialogConfig
+  ): Observable<DialogState<T>>
+  /**
+   * @deprecated Use `extras` instead of `showXButton`
+   */
+  openDialog<T>(
+    title: TranslationKey | null,
+    componentOrMessage: Type<any> | Type<DialogResult<T>> | Component<T> | TranslationKey | DialogMessage,
+    primaryButtonTranslationKeyOrDetails: TranslationKey | ButtonDialogButtonDetails,
+    secondaryButtonTranslationKeyOrDetails?: TranslationKey | ButtonDialogButtonDetails,
+    showXButton?: boolean
+  ): Observable<DialogState<T>>
+  openDialog<T>(
+    title: TranslationKey | null,
+    componentOrMessage: Type<any> | Type<DialogResult<T>> | Component<T> | TranslationKey | DialogMessage,
+    primaryButtonTranslationKeyOrDetails: TranslationKey | ButtonDialogButtonDetails,
+    secondaryButtonTranslationKeyOrDetails?: TranslationKey | ButtonDialogButtonDetails,
+    extrasOrShowXButton: PortalDialogConfig | boolean = {}
   ): Observable<DialogState<T>> {
+    const dialogOptions: PortalDialogConfig =
+      typeof extrasOrShowXButton === 'object'
+        ? extrasOrShowXButton
+        : {
+            showXButton: extrasOrShowXButton,
+          }
     const translateParams = this.prepareTitleForTranslation(title)
 
     const componentToRender: Component<any> = this.getComponentToRender(componentOrMessage)
@@ -386,7 +434,8 @@ export class PortalDialogService {
         return this.dialogService.open(ButtonDialogComponent, {
           header: dialogTitle,
           data: dynamicDialogDataConfig,
-          closable: showXButton && secondaryButtonTranslationKeyOrDetails !== undefined,
+          closable: dialogOptions.showXButton && secondaryButtonTranslationKeyOrDetails !== undefined,
+          ...dialogOptions,
         }).onClose
       })
     )